requirements.txt
pandas
streamlit
pillow
tabulate
plotly